Refactor: 하단 탭 코드 분리
