+++ WDG_iOS/View/ContentView.swift
@@ -118,53 +118,6 @@ struct ContentView: View {
     }
 }
 
-struct MyTabView: View {
-    @Binding var selectedTab: Int
-    @Binding var scrollProxy: ScrollViewProxy?
-    var namespace: Namespace.ID
-    var body: some View {
-        HStack {
-            Spacer()
-            Button(action: {
-                if self.selectedTab == 0 { scrollToTop() }
-                self.selectedTab = 0
-            }, label: {
-                Image(systemName: "list.bullet")
-                    .resizable()
-                    .frame(width: 30, height: 30)
-            })
-            Spacer()
-            Spacer()
-            Button(action: {
-                self.selectedTab = 1
-            }, label: {
-                Image(systemName: "square.and.pencil")
-                    .resizable()
-                    .frame(width: 30, height: 30)
-            })
-            Spacer()
-            Spacer()
-            Button(action: {
-                self.selectedTab = 2
-            }, label: {
-                Image(systemName: "person")
-                    .resizable()
-                    .frame(width: 30, height: 30)
-            })
-            Spacer()
-        }
-        .padding(.horizontal)
-        .frame(height: 80)
-        .background(Rectangle().foregroundColor(.white))
-        .foregroundColor(.black)
-    }
-    func scrollToTop() {
-        withAnimation {
-            scrollProxy?.scrollTo(namespace)
-        }
-    }
-}
-
 extension Binding where Value == Bool {
     /// A binding to the inverse of the bool value.
     var inverted: Binding<Bool> {
